% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor_adaptive.R
\name{check_args.nearest_neighbor_adaptive}
\alias{check_args.nearest_neighbor_adaptive}
\title{Check arguments of a model specification.}
\usage{
\method{check_args}{nearest_neighbor_adaptive}(object, call = rlang::caller_env())
}
\arguments{
\item{object}{A model specification.}

\item{call}{The environment used for error reporting.}
}
\value{
The model specification, invisibly.
}
\description{
Errors when main arguments are set that the selected engine
cannot use. Called by parsnip during fit().
}
\keyword{internal}
