% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{matrix_diagonal}
\alias{matrix_diagonal}
\title{Softening}
\usage{
matrix_diagonal(range = c(0, 2), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the defaults for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the transformed units.}

\item{trans}{A trans object from the scales package, such as
scales::log10_trans() or scales::reciprocal_trans(). If not provided, the
default is used, which matches the units used in range. If no
transformation, NULL.}
}
\value{
An S3 class of type quant_param from the dials package.
}
\description{
Scales the identity matrix added to the between class
covariance, which keeps the neighborhood from collapsing onto the class
boundary.
}
\details{
A value of 1 matches the publication. Of the tuning parameters,
this one usually has the smallest effect on performance.
}
\examples{
library(tidydann)

matrix_diagonal()
}
