% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tidy_sub_dann}
\alias{print.tidy_sub_dann}
\title{Print sub dann model.}
\usage{
\method{print}{tidy_sub_dann}(x, ...)
}
\arguments{
\item{x}{a dann model.}

\item{...}{arguments passed to other methods.}
}
\description{
Print sub dann model.
}
\examples{

library(tidymodels)
library(tidydann)

data("two_class_dat", package = "modeldata")
set.seed(1)
example_split <- initial_split(two_class_dat, prop = 0.99)
example_train <- training(example_split)
example_test <- testing(example_split)

model <- tidy_sub_dann() |>
  set_engine("sub_dann")

print(model)

}
