% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{sphere}
\alias{sphere}
\title{Sphere argument to ncoord}
\usage{
sphere(values = c("mcd", "mve", "classical", "none"))
}
\arguments{
\item{values}{A vector of candidate values. Any combination of "mcd",
"mve", "classical", and "none".}
}
\value{
An S3 class of type qual_param from the dials package.
}
\description{
Sphere argument to ncoord
}
\examples{
library(tidydann)

sphere()
}
