% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{tunable_tidy_sub_dann}
\alias{tunable_tidy_sub_dann}
\title{Declare tunable parameters}
\usage{
tunable_tidy_sub_dann(x, ...)
}
\arguments{
\item{x}{A model specification of type tidy_sub_dann
specification.}

\item{...}{Other arguments passed to methods.}
}
\value{
A tibble with a column for the parameter name, information on the default method for
generating a corresponding parameter object, the source of the parameter (e.g. "recipe", etc.),
and the component within the source.
}
\description{
Returns information on potential hyper-parameters that can be optimized.
}
