% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dann.R
\name{update.tidy_dann}
\alias{update.tidy_dann}
\title{Update method for tidy_dann}
\usage{
\method{update}{tidy_dann}(
  object,
  parameters = NULL,
  neighbors = NULL,
  neighborhood = NULL,
  matrix_diagonal = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{A model specification.}

\item{parameters}{A 1-row tibble or named list with main parameters to
update. Use either parameters or the main arguments directly when updating.
If the main arguments are used, these will supersede the values in
parameters. Also, using engine arguments in this object will result in an
error.}

\item{neighbors}{The number of data points used for final classification.}

\item{neighborhood}{The number of data points used to calculate between and
within class covariance.}

\item{matrix_diagonal}{Diagonal elements of a diagonal matrix. 1 is the
identity
matrix.}

\item{fresh}{A logical for whether the arguments should be modified
in-place or replaced wholesale.}

\item{...}{Not used for update().}
}
\value{
An updated parsnip spec for tidy_dann model.
}
\description{
Update method for tidy_dann
}
