% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sub_dann.R
\name{update.tidy_sub_dann}
\alias{update.tidy_sub_dann}
\title{Update method for tidy_sub_dann}
\usage{
\method{update}{tidy_sub_dann}(
  object,
  parameters = NULL,
  neighbors = NULL,
  neighborhood = NULL,
  epsilon = NULL,
  weighted = NULL,
  sphere = NULL,
  num_comp = NULL,
  fresh = FALSE,
  ...
)
}
\description{
Update method for tidy_sub_dann
}
