% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{neighborhood}
\alias{neighborhood}
\title{Neighborhood size}
\usage{
neighborhood(range = c(2L, dials::unknown()), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the defaults for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the transformed units.}

\item{trans}{A trans object from the scales package, such as
scales::log10_trans() or scales::reciprocal_trans().If not provided, the
default is used which matches the units used in range. If no transformation,
NULL.}
}
\description{
Number of data points used to calculate shape of the
neighborhood.
}
\details{
Use get_p from dials to finalize.
}
\examples{
library(dials)
library(tidydann)

data("two_class_dat", package = "modeldata")
neighborhood() |> get_p(two_class_dat[-1])
}
