% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{sphere}
\alias{sphere}
\title{Sphere argument to ncoord}
\usage{
sphere(values = c("mcd", "mve", "classical", "none"))
}
\arguments{
\item{values}{A four-element vector containing "mcd", "mve", "classical", and "none".}
}
\description{
Sphere argument to ncoord
}
\examples{
library(tidydann)

sphere()
}
