% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_dann.R
\name{update.tidy_dann}
\alias{update.tidy_dann}
\title{Update method for tidy_dann}
\usage{
\method{update}{tidy_dann}(
  object,
  parameters = NULL,
  neighbors = NULL,
  neighborhood = NULL,
  epsilon = NULL,
  fresh = FALSE,
  ...
)
}
\description{
Update method for tidy_dann
}
