% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{weighted}
\alias{weighted}
\title{Weighted argument to ncoord}
\usage{
weighted(values = c(FALSE, TRUE))
}
\arguments{
\item{values}{A one-element vector containing FALSE or TRUE.}
}
\value{
An S3 class of type qual_param from the dials package.
}
\description{
Weighted argument to ncoord
}
\examples{
library(tidydann)

weighted()
}
