% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunable.R
\name{weighted}
\alias{weighted}
\title{Weighted argument to ncoord}
\usage{
weighted(values = c(FALSE, TRUE))
}
\arguments{
\item{values}{A two-element vector containing FALSE and TRUE.}
}
\description{
Weighted argument to ncoord
}
\examples{
library(tidydann)

weighted()
}
